refactor(jules): follow idiomatic Go patterns (#39)
* refactor: clean up non-idiomatic Go patterns, formatting, and styles

- Refactor `internal/core/bridge_test.go` to use standard library `slices.Equal` instead of custom comparison.
- Update `extractPlainTextBody` in `internal/email/adapter_fetch.go` to use idiomatic EOF checks.
- Simplify transaction handling in `WithTx` using standard `defer func() { _ = sqlTx.Rollback() }()` to satisfy `errcheck`.
- Scan directly into `uint32` inside `Cursor` and `SetCursor` rather than casting.
- Use `errors.Is` for `pflag.ErrHelp` comparisons.
- Prevent variable shadowing of `err` inside database transaction closures.


---------

+++ internal/store/sqlite/email.go
@@ -7,7 +7,7 @@ import (
 
 // Cursor returns the current UID and UIDValidity values of the local IMAP client.
 func (t *Tx) Cursor(ctx context.Context, username, mailbox string) (uint32, uint32, error) {
-	var uid, uidValidity int
+	var uid, uidValidity uint32
 	err := t.Tx.QueryRowContext(
 		ctx,
 		`SELECT last_seen_uid, uid_validity
@@ -19,7 +19,7 @@ func (t *Tx) Cursor(ctx context.Context, username, mailbox string) (uint32, uint
 	if err != nil {
 		return 0, 0, err
 	}
-	return uint32(uid), uint32(uidValidity), nil
+	return uid, uidValidity, nil
 }
 
 // SetCursor updates the current UID and UIDValidity values of the local IMAP client.
